Add acceptance test fix code

+++ lib/Characters.cpp
@@ -2,8 +2,8 @@
 
 #include <algorithm>
 #include <cctype>
+#include <iostream>
 #include <numeric>
-
 using namespace swo::chars;
 
 void Characters::collect(char input) { data.push_back(input); }
@@ -12,5 +12,5 @@ double Characters::letterFrequency() const {
   const auto letters =
       std::count_if(std::begin(data), std::end(data),
                     [](unsigned char ch) { return std::isalpha(ch); });
-  return letters / count();
+  return (double)letters / count();
 }
